alpha-pod/stream-metrics/requirements.txt
kafka-python==2.0.2
prometheus-api-client==0.4.2